docs(): changed API_URL env to localhost
The current `dot_env_example` env `API_URL` contains as host 0.0.0.0.
This will likely not work as this is a binding the server to all
interfaces. The worker is using the container name as host so this
"should" not break anything. This commit prevents the error message when
connecting to the prefect UI with localhost:4200.